# Optimize navigation