--- a/phpstan-baseline.neon
+++ b/phpstan-baseline.neon
@@ -9891,13 +9891,8 @@ parameters:
 			path: src/Navigation/Nodes/NodeDatabase.php
 
 		-
-			message: "#^Parameter \\#2 \\$callback of function usort expects callable\\(mixed, mixed\\)\\: int, 'strnatcasecmp' given\\.$#"
-			count: 3
-			path: src/Navigation/Nodes/NodeDatabase.php
-
-		-
-			message: "#^Parameter \\#2 \\$callback of function usort expects callable\\(string\\|null, string\\|null\\)\\: int, 'strnatcasecmp' given\\.$#"
-			count: 3
+			message: "#^Parameter \\#2 \\$callback of function usort expects callable\\(mixed, mixed\\)\\: int, Closure\\(string, string\\)\\: int\\<\\-1, 1\\> given\\.$#"
+			count: 1
 			path: src/Navigation/Nodes/NodeDatabase.php
 
 		-