… psalm-baseline.xml
… tests/classes/Controllers/NavigationControllerTest.php
… tests/classes/Dbal/MysqliResultTest.php
… tests/classes/Navigation/Nodes/ObjectFetcherTest.php
… tests/classes/Stubs/DummyResult.php