@liviuconcioiu 2024-02-03
Awesome work! On my side, from 1.4s to 0.7s.